requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
cryptography==39.0.0
csscompressor==0.9.5
defusedxml==0.7.1
Django==4.1.5
django-filter==22.1
django-modeltranslation==0.18.9
django-rosetta==0.9.8
djangorestframework==3.14.0
ginger==0.0.4
idna==3.4
importlib-metadata==6.0.0
Jinja2==3.1.2
jsmin==3.0.1
libsass==0.22.0
Markdown==3.4.1
MarkupSafe==2.1.2
oauthlib==3.2.2
Pillow==9.4.0
placeholder==1.4
polib==1.1.1
psycopg2-binary==2.9.5
pycparser==2.21
PyJWT==2.6.0
python3-openid==3.2.0
pytz==2022.7.1
PyYAML==6.0
requests==2.28.2
requests-oauthlib==1.3.1
rosetta==0.3
social-auth-app-django==5.0.0
social-auth-core==4.3.0
sqlparse==0.4.3
translation==1.0.5
translations==0.1.3
typing-extensions==4.4.0
urllib3==1.26.14
watchdog==2.2.1
zipp==3.13.0
